// web/src/lib/generated-api/song-api/models/song.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * SongService | v1
 * OpenAPI spec version: 1.0.0
 */
import type { SongAuthor } from './songAuthor';
import type { SongLyrics } from './songLyrics';
import type { SongChords } from './songChords';
import type { SongKey } from './songKey';
import type { SongBpm } from './songBpm';
import type { SongUpdatedAt } from './songUpdatedAt';
import type { SongUpdatedBy } from './songUpdatedBy';

export interface Song {
  id: string;
  bandId: string;
  title: string;
  author?: SongAuthor;
  lyrics?: SongLyrics;
  chords?: SongChords;
  key?: SongKey;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  bpm?: SongBpm;
  createdAt: string;
  createdBy: string;
  updatedAt?: SongUpdatedAt;
  updatedBy?: SongUpdatedBy;
}